test: cover canonical plugin config detection
